+++ subs/curator/src/Curator/StackageConstraints.hs
@@ -54,7 +54,8 @@ data SC = SC
   , scSkippedHaddocks :: !(Set PackageName)
   -- ^ Sometimes Haddock is really flaky
 
-  -- FIXME deal with all of the github-users and ping logic
+  , scGithubUsers :: !(Map Text (Set Text))
+  -- ^ Mapping from Github org to users who will receive pings
 
   , scTellMeWhenItsReleased :: !(Map PackageName Version)
 
@@ -81,6 +82,8 @@ instance FromJSON SC where
     scExpectedTestFailures <- Set.map unCabalString <$> o .: "expected-test-failures"
     scExpectedBenchmarkFailures <- Set.map unCabalString <$> o .: "expected-benchmark-failures"
     scExpectedHaddockFailures <- Set.map unCabalString <$> o .: "expected-haddock-failures"
+
+    scGithubUsers <- o .: "github-users"
 
     scHide <- Set.map unCabalString <$> o .: "hide"
     scNoRevisions <- Set.map unCabalString <$> o .: "no-revisions"
@@ -138,6 +141,7 @@ convert sc0 = do
   pure Constraints
     { consGhcVersion = scGhcVersion sc1
     , consPackages = packages
+    , consGithubUsers = scGithubUsers sc1
     }
   where
     go :: (SC, Map PackageName PackageConstraints, [String])