+++ subs/pantry/src/Pantry.hs
@@ -20,6 +20,7 @@ module Pantry
   , PantryApp
   , runPantryApp
   , runPantryAppClean
+  , hpackExecutableL
 
     -- * Types
 
@@ -147,6 +148,7 @@ module Pantry
   , loadCabalFileRawImmutable
   , loadCabalFileImmutable
   , loadCabalFilePath
+  , findOrGenerateCabalFile
   , PrintWarnings (..)
 
     -- * Hackage index
@@ -158,6 +160,7 @@ module Pantry
   , getLatestHackageLocation
   , getLatestHackageRevision
   , getHackageTypoCorrections
+  , loadGlobalHints
   ) where
 
 import RIO
@@ -182,6 +185,8 @@ import qualified Distribution.PackageDescription as D
 import Distribution.Parsec.Common (PWarning (..), showPos)
 import qualified Hpack
 import qualified Hpack.Config as Hpack
+import Network.HTTP.Download
+import RIO.PrettyPrint
 import RIO.Process
 import RIO.Directory (getAppUserDataDirectory)
 import qualified Data.Yaml as Yaml
@@ -294,7 +299,7 @@ getLatestHackageLocation name preferred = do
 --
 -- @since 0.1.0.0
 getLatestHackageRevision
-  :: (HasPantryConfig env, HasLogFunc env)
+  :: (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
   => PackageName -- ^ package name
   -> Version
   -> RIO env (Maybe (Revision, BlobKey, TreeKey))
@@ -647,7 +652,9 @@ loadCabalFileBytes (PLIHackage pident cfHash _mtree) = getHackageCabalFile (pirF
 loadCabalFileBytes pl = do
   package <- loadPackage pl
   let sfp = cabalFileName $ pkgName $ packageIdent package
-      TreeEntry cabalBlobKey _ft = packageCabalEntry package
+  cabalBlobKey <- case (packageCabalEntry package) of
+                       PCHpack pcHpack -> pure $ teBlob . phGenerated $ pcHpack
+                       PCCabalFile (TreeEntry blobKey _) -> pure blobKey
   mbs <- withStorage $ loadBlob cabalBlobKey
   case mbs of
     Nothing -> do
@@ -669,7 +676,9 @@ loadRawCabalFileBytes (RPLIHackage pir _mtree) = getHackageCabalFile pir
 loadRawCabalFileBytes pl = do
   package <- loadPackageRaw pl
   let sfp = cabalFileName $ pkgName $ packageIdent package
-      TreeEntry cabalBlobKey _ft = packageCabalEntry package
+      TreeEntry cabalBlobKey _ft = case packageCabalEntry package of
+                                     PCCabalFile cabalTE -> cabalTE
+                                     PCHpack hpackCE -> phGenerated hpackCE
   mbs <- withStorage $ loadBlob cabalBlobKey
   case mbs of
     Nothing -> do
@@ -754,7 +763,9 @@ completePM plOrig rpm@(RawPackageMetadata mn mv mtk mc)
       let pm = PackageMetadata
             { pmIdent = packageIdent package
             , pmTreeKey = packageTreeKey package
-            , pmCabal = teBlob $ packageCabalEntry package
+            , pmCabal = teBlob $ case packageCabalEntry package of
+                                   PCCabalFile cfile -> cfile
+                                   PCHpack hfile -> phGenerated hfile
             }
 
           isSame x (Just y) = x == y
@@ -1368,6 +1379,9 @@ data PantryApp = PantryApp
 simpleAppL :: Lens' PantryApp SimpleApp
 simpleAppL = lens paSimpleApp (\x y -> x { paSimpleApp = y })
 
+hpackExecutableL :: Lens' PantryConfig HpackExecutable
+hpackExecutableL k pconfig = fmap (\hpExe -> pconfig { pcHpackExecutable = hpExe }) (k (pcHpackExecutable pconfig))
+
 instance HasLogFunc PantryApp where
   logFuncL = simpleAppL.logFuncL
 instance HasPantryConfig PantryApp where
@@ -1418,3 +1432,40 @@ runPantryAppClean f = liftIO $ withSystemTempDirectory "pantry-clean" $ \dir -> 
           , paPantryConfig = pc
           }
         f
+
+-- | Load the global hints from Github.
+--
+-- @since 0.1.0.0
+loadGlobalHints
+  :: HasTerm env
+  => Path Abs File -- ^ local cached file location
+  -> WantedCompiler
+  -> RIO env (Maybe (Map PackageName Version))
+loadGlobalHints dest wc =
+    inner False
+  where
+    inner alreadyDownloaded = do
+      req <- parseRequest "https://raw.githubusercontent.com/fpco/stackage-content/master/stack/global-hints.yaml"
+      downloaded <- download req dest
+      eres <- tryAny inner2
+      mres <-
+        case eres of
+          Left e -> Nothing <$ logError ("Error when parsing global hints: " <> displayShow e)
+          Right x -> pure x
+      case mres of
+        Nothing | not alreadyDownloaded && not downloaded -> do
+          logInfo $
+            "Could not find local global hints for " <>
+            RIO.display wc <>
+            ", forcing a redownload"
+          x <- redownload req dest
+          if x
+            then inner True
+            else do
+              logInfo "Redownload didn't happen"
+              pure Nothing
+        _ -> pure mres
+
+    inner2 = liftIO
+           $ Map.lookup wc . fmap (fmap unCabalString . unCabalStringMap)
+         <$> Yaml.decodeFileThrow (toFilePath dest)