+++ subs/pantry/src/Pantry/Storage.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
@@ -26,8 +28,10 @@ module Pantry.Storage
   , storeCacheUpdate
   , storeHackageTarballInfo
   , loadHackageTarballInfo
+  , getHPackBlobKeyById
   , storeTree
   , loadTree
+  , storeHPack
   , loadPackageById
   , getTreeForKey
   , storeHackageTree
@@ -40,6 +44,8 @@ module Pantry.Storage
   , storePreferredVersion
   , loadPreferredVersion
   , sinkHackagePackageNames
+  , loadCabalBlobKey
+  , hpackToCabal
   , countHackageCabals
 
     -- avoid warnings
@@ -57,21 +63,28 @@ module Pantry.Storage
   ) where
 
 import RIO hiding (FilePath)
+import RIO.Process
 import qualified RIO.ByteString as B
 import qualified Pantry.Types as P
+import qualified RIO.List as List
+import qualified RIO.FilePath as FilePath
+import RIO.FilePath ((</>), takeDirectory)
+import RIO.Directory (createDirectoryIfMissing, setPermissions, getPermissions, setOwnerExecutable)
 import Database.Persist
 import Database.Persist.Sqlite
 import Database.Persist.TH
 import RIO.Orphans ()
 import qualified Pantry.SHA256 as SHA256
 import qualified RIO.Map as Map
+import qualified RIO.Text as T
 import RIO.Time (UTCTime, getCurrentTime)
-import Path (Path, Abs, File, toFilePath, parent)
-import Path.IO (ensureDir)
+import Path (Path, Abs, File, Dir, toFilePath, parent, filename, parseAbsDir, fromAbsFile, fromRelFile)
+import Path.IO (ensureDir, listDir, createTempDir, getTempDir, removeDirRecur)
 import Data.Pool (destroyAllResources)
+import Pantry.HPack (hpackVersion, hpack)
 import Conduit
 import Data.Acquire (with)
-import Pantry.Types (PackageNameP (..), VersionP (..), SHA256, FileSize (..), FileType (..), HasPantryConfig, BlobKey, Repo (..), TreeKey, SafeFilePath, Revision (..), Package (..))
+import Pantry.Types (PackageNameP (..), VersionP (..), SHA256, FileSize (..), FileType (..), HasPantryConfig, BlobKey, Repo (..), TreeKey, SafeFilePath, Revision (..), Package (..), PantryException (MigrationFailure))
 
 share [mkPersist sqlSettings, mkMigrate "migrateAll"] [persistLowerCase|
 -- Raw blobs
@@ -146,11 +159,27 @@ CacheUpdate
 -- table.
 Tree
     key BlobId
-    cabal BlobId
+
+    -- If the treeCabal field is Nothing, it means the Haskell package
+    -- doesn't have a corresponding cabal file for it. This may be the case
+    -- for haskell package referenced by git repository with only a hpack file.
+    cabal BlobId Maybe
     cabalType FileType
     name PackageNameId
     version VersionId
     UniqueTree key
+
+HPack
+   tree TreeId
+
+   -- hpack version used for generating this cabal file
+   version VersionId
+
+   -- Generated cabal file for the given tree and hpack version
+   cabalBlob BlobId
+   cabalPath FilePathId
+
+   UniqueHPack tree version
 
 -- An individual file within a Tree.
 TreeEntry
@@ -186,15 +215,17 @@ initStorage
 initStorage fp inner = do
   ensureDir $ parent fp
   bracket
-    (createSqlitePoolFromInfo sqinfo 1)
+    (createSqlitePoolFromInfo (sqinfo False) 1)
     (liftIO . destroyAllResources) $ \pool -> do
-
-    migrates <- runSqlPool (runMigrationSilent migrateAll) pool
-    forM_ migrates $ \mig -> logDebug $ "Migration output: " <> display mig
-    inner (P.Storage pool)
+    migrates <- wrapMigrationFailure $ runSqlPool (runMigrationSilent migrateAll) pool
+    forM_ migrates $ \mig -> logDebug $ "Migration executed: " <> display mig
+  bracket
+    (createSqlitePoolFromInfo (sqinfo True) 1)
+    (liftIO . destroyAllResources) $ \pool -> inner (P.Storage pool)
   where
-    sqinfo = set extraPragmas ["PRAGMA busy_timeout=2000;"]
-           $ set fkEnabled True
+    wrapMigrationFailure = handle (throwIO . MigrationFailure fp)
+    sqinfo fk = set extraPragmas ["PRAGMA busy_timeout=2000;"]
+           $ set fkEnabled fk
            $ mkSqliteConnectionInfo (fromString $ toFilePath fp)
 
 withStorage
@@ -272,6 +303,7 @@ loadBlobById bid = do
   case mbt of
     Nothing -> error "loadBlobById: ID doesn't exist in database"
     Just bt -> pure $ blobContents bt
+
 
 getBlobKey
   :: (HasPantryConfig env, HasLogFunc env)
@@ -437,30 +469,120 @@ loadHackageTarballInfo name version = do
   where
     go (Entity _ ht) = (hackageTarballSha ht, hackageTarballSize ht)
 
+storeCabalFile ::
+       (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => ByteString
+    -> P.PackageName
+    -> ReaderT SqlBackend (RIO env) BlobId
+storeCabalFile cabalBS pkgName = do
+    (bid, _) <- storeBlob cabalBS
+    let cabalFile = P.cabalFileName pkgName
+    _ <- insertBy FilePath {filePathPath = cabalFile}
+    return bid
+
+loadFilePath ::
+       (HasPantryConfig env, HasLogFunc env)
+    => SafeFilePath
+    -> ReaderT SqlBackend (RIO env) (Entity FilePath)
+loadFilePath path = do
+    fp <- getBy $ UniqueSfp path
+    case fp of
+        Nothing ->
+            error $
+            "loadFilePath: No row found for " <>
+            (T.unpack $ P.unSafeFilePath path)
+        Just record -> return record
+
+loadHPackTreeEntity :: (HasPantryConfig env, HasLogFunc env) => TreeId -> ReaderT SqlBackend (RIO env) (Entity TreeEntry)
+loadHPackTreeEntity tid = do
+  filepath <- loadFilePath P.hpackSafeFilePath
+  let filePathId :: FilePathId = entityKey filepath
+  hpackTreeEntry <-
+      selectFirst [TreeEntryTree ==. tid, TreeEntryPath ==. filePathId] []
+  hpackEntity <-
+      case hpackTreeEntry of
+        Nothing ->
+            error $ "loadHPackTreeEntity: No package.yaml file found in TreeEntry for TreeId:  " ++ (show tid)
+        Just record -> return record
+  return hpackEntity
+
+storeHPack ::
+       (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => P.RawPackageLocationImmutable
+    -> TreeId
+    -> ReaderT SqlBackend (RIO env) (Key HPack)
+storeHPack rpli tid = do
+    vid <- hpackVersionId
+    hpackRecord <- getBy (UniqueHPack tid vid)
+    case hpackRecord of
+      Nothing -> generateHPack rpli tid vid
+      Just record -> return $ entityKey record
+
+loadCabalBlobKey :: (HasPantryConfig env, HasLogFunc env) => HPackId -> ReaderT SqlBackend (RIO env) BlobKey
+loadCabalBlobKey hpackId = do
+  hpackRecord <- getJust hpackId
+  getBlobKey $ hPackCabalBlob hpackRecord
+
+generateHPack ::
+       (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => P.RawPackageLocationImmutable -- ^ for exceptions
+    -> TreeId
+    -> VersionId
+    -> ReaderT SqlBackend (RIO env) (Key HPack)
+generateHPack rpli tid vid = do
+    tree <- getTree tid
+    (pkgName, cabalBS) <- hpackToCabalS rpli tree
+    bid <- storeCabalFile cabalBS pkgName
+    let cabalFile = P.cabalFileName pkgName
+    fid <- insertBy FilePath {filePathPath = cabalFile}
+    let hpackRecord =
+            HPack
+                { hPackTree = tid
+                , hPackVersion = vid
+                , hPackCabalBlob = bid
+                , hPackCabalPath = either entityKey id fid
+                }
+    either entityKey id <$> insertBy hpackRecord
+
+
+hpackVersionId ::
+       (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => ReaderT SqlBackend (RIO env) VersionId
+hpackVersionId = do
+    hpackSoftwareVersion <- lift $ hpackVersion
+    fmap (either entityKey id) $
+      insertBy $
+      Version {versionVersion = P.VersionP hpackSoftwareVersion}
+
 storeTree
-  :: (HasPantryConfig env, HasLogFunc env)
-  => P.PackageIdentifier
+  :: (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+  => P.RawPackageLocationImmutable -- ^ for exceptions
+  -> P.PackageIdentifier
   -> P.Tree
-  -> P.TreeEntry
-  -- ^ cabal file
+  -> P.BuildFile
   -> ReaderT SqlBackend (RIO env) (TreeId, P.TreeKey)
-storeTree (P.PackageIdentifier name version) tree@(P.TreeMap m) (P.TreeEntry (P.BlobKey cabal _) cabalType) = do
+storeTree rpli (P.PackageIdentifier name version) tree@(P.TreeMap m) buildFile = do
   (bid, blobKey) <- storeBlob $ P.renderTree tree
-  mcabalid <- loadBlobBySHA cabal
-  cabalid <-
-    case mcabalid of
-      Just cabalid -> pure cabalid
-      Nothing -> error $ "storeTree: cabal BlobKey not found: " ++ show (tree, cabal)
+  (cabalid, ftype) <- case buildFile of
+                P.BFHpack (P.TreeEntry _ ftype) -> pure (Nothing, ftype)
+                P.BFCabal _ (P.TreeEntry (P.BlobKey btypeSha _) ftype) -> do
+                               buildTypeid <- loadBlobBySHA btypeSha
+                               buildid <-
+                                   case buildTypeid of
+                                     Just buildId -> pure buildId
+                                     Nothing -> error $ "storeTree: " ++ (show buildFile) ++ " BlobKey not found: " ++ show (tree, btypeSha)
+                               return (Just buildid, ftype)
   nameid <- getPackageNameId name
   versionid <- getVersionId version
   etid <- insertBy Tree
     { treeKey = bid
     , treeCabal = cabalid
-    , treeCabalType = cabalType
+    , treeCabalType = ftype
     , treeName = nameid
     , treeVersion = versionid
     }
-  case etid of
+
+  (tid, pTreeKey) <- case etid of
     Left (Entity tid _) -> pure (tid, P.TreeKey blobKey) -- already in database, assume it matches
     Right tid -> do
       for_ (Map.toList m) $ \(sfp, P.TreeEntry blobKey' ft) -> do
@@ -477,6 +599,22 @@ storeTree (P.PackageIdentifier name version) tree@(P.TreeMap m) (P.TreeEntry (P.
           , treeEntryType = ft
           }
       pure (tid, P.TreeKey blobKey)
+  case buildFile of
+    P.BFHpack _ -> storeHPack rpli tid >> return ()
+    P.BFCabal _ _ -> return ()
+  return (tid, pTreeKey)
+
+getTree :: (HasPantryConfig env, HasLogFunc env)
+  => TreeId
+  -> ReaderT SqlBackend (RIO env) P.Tree
+getTree tid = do
+  (mts :: Maybe Tree) <- get tid
+  ts <-
+      case mts of
+        Nothing ->
+            error $ "getTree: invalid foreign key " ++ show tid
+        Just ts -> pure ts
+  loadTreeByEnt $ Entity tid ts
 
 loadTree
   :: (HasPantryConfig env, HasLogFunc env)
@@ -498,40 +636,102 @@ getTreeForKey (P.TreeKey key) = do
     Nothing -> pure Nothing
     Just bid -> getBy $ UniqueTree bid
 
-loadPackageById
-  :: (HasPantryConfig env, HasLogFunc env)
-  => TreeId
-  -> ReaderT SqlBackend (RIO env) Package
-loadPackageById tid = do
-  mts <- get tid
-  ts <-
-    case mts of
-      Nothing -> error $ "loadPackageById: invalid foreign key " ++ show tid
-      Just ts -> pure ts
-  tree <- loadTreeByEnt $ Entity tid ts
-  key <- getBlobKey $ treeKey ts
+loadPackageById ::
+       (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => P.RawPackageLocationImmutable -- ^ for exceptions
+    -> TreeId
+    -> ReaderT SqlBackend (RIO env) Package
+loadPackageById rpli tid = do
+    (mts :: Maybe Tree) <- get tid
+    ts <-
+        case mts of
+            Nothing ->
+                error $ "loadPackageById: invalid foreign key " ++ show tid
+            Just ts -> pure ts
+    (tree :: P.Tree) <- loadTreeByEnt $ Entity tid ts
+    (blobKey :: BlobKey) <- getBlobKey $ treeKey ts
+    (mname :: Maybe PackageName) <- get $ treeName ts
+    name <-
+        case mname of
+            Nothing ->
+                error $
+                "loadPackageByid: invalid foreign key " ++ show (treeName ts)
+            Just (PackageName (P.PackageNameP name)) -> pure name
+    mversion <- get $ treeVersion ts
+    version <-
+        case mversion of
+            Nothing ->
+                error $
+                "loadPackageByid: invalid foreign key " ++ show (treeVersion ts)
+            Just (Version (P.VersionP version)) -> pure version
+    let ident = P.PackageIdentifier name version
+    (pentry, mtree) <-
+        case (treeCabal ts) of
+            Just keyBlob -> do
+                cabalKey <- getBlobKey keyBlob
+                return
+                    ( P.PCCabalFile $ P.TreeEntry cabalKey (treeCabalType ts)
+                    , tree)
+            Nothing -> do
+                hpackVid <- hpackVersionId
+                hpackEntity <- getBy (UniqueHPack tid hpackVid)
+                let (P.TreeMap tmap) = tree
+                    cabalFile = P.cabalFileName name
+                case hpackEntity of
+                    Nothing
+                        -- This case will happen when you either
+                        -- update stack with a new hpack version or
+                        -- use different hpack version via
+                        -- --with-hpack option.
+                     -> do
+                        (hpackId :: HPackId) <- storeHPack rpli tid
+                        hpackRecord <- getJust hpackId
+                        getHPackCabalFile hpackRecord ts tmap cabalFile
+                    Just (Entity _ item) ->
+                        getHPackCabalFile item ts tmap cabalFile
+    pure
+        Package
+            { packageTreeKey = P.TreeKey blobKey
+            , packageTree = mtree
+            , packageCabalEntry = pentry
+            , packageIdent = ident
+            }
 
-  mname <- get $ treeName ts
-  name <-
-    case mname of
-      Nothing -> error $ "loadPackageByid: invalid foreign key " ++ show (treeName ts)
-      Just (PackageName (P.PackageNameP name)) -> pure name
+getHPackBlobKey :: (HasPantryConfig env, HasLogFunc env) => HPack -> ReaderT SqlBackend (RIO env) BlobKey
+getHPackBlobKey hpackRecord = do
+  let treeId = hPackTree hpackRecord
+  hpackEntity <- loadHPackTreeEntity treeId
+  getBlobKey (treeEntryBlob $ entityVal hpackEntity)
 
-  mversion <- get $ treeVersion ts
-  version <-
-    case mversion of
-      Nothing -> error $ "loadPackageByid: invalid foreign key " ++ show (treeVersion ts)
-      Just (Version (P.VersionP version)) -> pure version
+getHPackBlobKeyById :: (HasPantryConfig env, HasLogFunc env) => HPackId -> ReaderT SqlBackend (RIO env) BlobKey
+getHPackBlobKeyById hpackId = do
+  hpackRecord <- getJust hpackId
+  getHPackBlobKey hpackRecord
 
-  cabalKey <- getBlobKey $ treeCabal ts
-  let ident = P.PackageIdentifier name version
-  let cabalEntry = P.TreeEntry cabalKey (treeCabalType ts)
-  pure Package
-    { packageTreeKey = P.TreeKey key
-    , packageTree = tree
-    , packageCabalEntry = cabalEntry
-    , packageIdent = ident
-    }
+
+getHPackCabalFile ::
+       (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => HPack
+    -> Tree
+    -> Map SafeFilePath P.TreeEntry
+    -> SafeFilePath
+    -> ReaderT SqlBackend (RIO env) (P.PackageCabal, P.Tree)
+getHPackCabalFile hpackRecord ts tmap cabalFile = do
+    cabalKey <- getBlobKey (hPackCabalBlob hpackRecord)
+    hpackKey <- getHPackBlobKey hpackRecord
+    hpackSoftwareVersion <- lift hpackVersion
+    let fileType = treeCabalType ts
+        cbTreeEntry = P.TreeEntry cabalKey fileType
+        hpackTreeEntry = P.TreeEntry hpackKey fileType
+        tree = P.TreeMap $ Map.insert cabalFile cbTreeEntry tmap
+    return $
+        ( P.PCHpack $
+          P.PHpack
+              { P.phOriginal = hpackTreeEntry
+              , P.phGenerated = cbTreeEntry
+              , P.phVersion = hpackSoftwareVersion
+              }
+        , tree)
 
 loadTreeByEnt
   :: (HasPantryConfig env, HasLogFunc env)
@@ -596,12 +796,13 @@ loadHackageTreeKey name ver sha = do
       pure $ Just $ P.TreeKey $ P.BlobKey treesha size
 
 loadHackageTree
-  :: (HasPantryConfig env, HasLogFunc env)
-  => P.PackageName
+  :: (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+  => P.RawPackageLocationImmutable -- ^ for exceptions
+  -> P.PackageName
   -> P.Version
   -> BlobId
   -> ReaderT SqlBackend (RIO env) (Maybe Package)
-loadHackageTree name ver bid = do
+loadHackageTree rpli name ver bid = do
   nameid <- getPackageNameId name
   versionid <- getVersionId ver
   ment <- selectFirst
@@ -616,7 +817,7 @@ loadHackageTree name ver bid = do
     Just (Entity _ hc) ->
       case hackageCabalTree hc of
         Nothing -> assert False $ pure Nothing
-        Just tid -> Just <$> loadPackageById tid
+        Just tid -> Just <$> loadPackageById rpli tid
 
 storeArchiveCache
   :: (HasPantryConfig env, HasLogFunc env)
@@ -728,6 +929,87 @@ sinkHackagePackageNames predicate sink = do
     checkOnHackage nameid = do
       cnt <- count [HackageCabalName ==. nameid]
       pure $ cnt > 0
+
+-- | Get the filename for the cabal file in the given directory.
+--
+-- If no .cabal file is present, or more than one is present, an exception is
+-- thrown via 'throwM'.
+--
+-- If the directory contains a file named package.yaml, hpack is used to
+-- generate a .cabal file from it.
+findOrGenerateCabalFile
+    :: forall env. (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+    => Path Abs Dir -- ^ package directory
+    -> RIO env (P.PackageName, Path Abs File)
+findOrGenerateCabalFile pkgDir = do
+    hpack pkgDir
+    files <- filter (flip hasExtension "cabal" . toFilePath) . snd
+         <$> listDir pkgDir
+    -- If there are multiple files, ignore files that start with
+    -- ".". On unixlike environments these are hidden, and this
+    -- character is not valid in package names. The main goal is
+    -- to ignore emacs lock files - see
+    -- https://github.com/commercialhaskell/stack/issues/1897.
+    let isHidden ('.':_) = True
+        isHidden _ = False
+    case filter (not . isHidden . fromRelFile . filename) files of
+        [] -> throwIO $ P.NoCabalFileFound pkgDir
+        [x] -> maybe
+          (throwIO $ P.InvalidCabalFilePath x)
+          (\pn -> pure $ (pn, x)) $
+            List.stripSuffix ".cabal" (toFilePath (filename x)) >>=
+            P.parsePackageName
+        _:_ -> throwIO $ P.MultipleCabalFilesFound pkgDir files
+      where hasExtension fp x = FilePath.takeExtension fp == "." ++ x
+
+-- | Similar to 'hpackToCabal' but doesn't require a new connection to database.
+hpackToCabalS :: (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+              => P.RawPackageLocationImmutable -- ^ for exceptions
+              -> P.Tree
+              -> ReaderT SqlBackend (RIO env) (P.PackageName, ByteString)
+hpackToCabalS rpli tree = do
+  tmpDir <- lift $ do
+              tdir <- getTempDir
+              createTempDir tdir "hpack-pkg-dir"
+  unpackTreeToDir rpli tmpDir tree
+  (packageName, cfile) <- lift $ findOrGenerateCabalFile tmpDir
+  !bs <- lift $ B.readFile (fromAbsFile cfile)
+  lift $ removeDirRecur tmpDir
+  return $ (packageName, bs)
+
+hpackToCabal :: (HasPantryConfig env, HasLogFunc env, HasProcessContext env)
+           => P.RawPackageLocationImmutable -- ^ for exceptions
+           -> P.Tree
+           -> RIO env (P.PackageName, ByteString)
+hpackToCabal rpli tree = withSystemTempDirectory "hpack-pkg-dir" $ \tmpdir -> do
+               tdir <- parseAbsDir tmpdir
+               withStorage $ unpackTreeToDir rpli tdir tree
+               (packageName, cfile) <- findOrGenerateCabalFile tdir
+               bs <- B.readFile (fromAbsFile cfile)
+               return (packageName, bs)
+
+unpackTreeToDir
+  :: (HasPantryConfig env, HasLogFunc env)
+  => P.RawPackageLocationImmutable -- ^ for exceptions
+  -> Path Abs Dir -- ^ dest dir, will be created if necessary
+  -> P.Tree
+  -> ReaderT SqlBackend (RIO env) ()
+unpackTreeToDir rpli (toFilePath -> dir) (P.TreeMap m) = do
+  for_  (Map.toList m) $ \(sfp, P.TreeEntry blobKey ft) -> do
+    let dest = dir </> T.unpack (P.unSafeFilePath sfp)
+    createDirectoryIfMissing True $ takeDirectory dest
+    mbs <- loadBlob blobKey
+    case mbs of
+      Nothing -> do
+        -- TODO when we have pantry wire stuff, try downloading
+        throwIO $ P.TreeReferencesMissingBlob rpli sfp blobKey
+      Just bs -> do
+        B.writeFile dest bs
+        case ft of
+          FTNormal -> pure ()
+          FTExecutable -> liftIO $ do
+            perms <- getPermissions dest
+            setPermissions dest $ setOwnerExecutable True perms
 
 countHackageCabals
   :: (HasPantryConfig env, HasLogFunc env)